+++ app/src/main/java/com/srm/srmapp/data/models/Food.kt
@@ -8,7 +8,7 @@ data class Food(
     val foodId: Int = -1,
     val name: String,
     val units: String,
-    val stockCount: Int = 0,
+    val stockCount: Float = 0f,
     val stockList: MutableList<Stock> = mutableListOf(),
 ) : GetId {
     override fun getId(): Int {
